move reducers/authReducers.js to auth/authReducers.js
